Move resource methods from TestHelper to ResourceHelper
